requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==3.2.0
email-validator
pydantic==1.10.13
requests

